+++ packages/pluggableWidgets/image-cropper-web/CHANGELOG.md
@@ -11,3 +11,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 - Initial release of Image cropper widget.
+- Custom aspect ratio width and height can be set with an expression or bound to an attribute.